api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
